+++ FuzzySharp.Benchmarks/BenchmarkAll.cs
@@ -200,7 +200,7 @@ public class BenchmarkAll
     [Benchmark]
     public int LevenshteinDistance()
     {
-        return NewLevenshtein.Distance("chicago cubs vs new york mets".AsSpan(), "new york mets vs chicago cubs".AsSpan());
+        return Levenshtein.Distance("chicago cubs vs new york mets".AsSpan(), "new york mets vs chicago cubs".AsSpan());
     }
 
     [Benchmark]
