+++ FuzzySharp/Delegates.cs
@@ -2,4 +2,5 @@
 
 namespace Raffinert.FuzzySharp;
 
+public delegate int Scorer(string input1, string input2);
 public delegate void Processor<T>(ref ReadOnlySpan<T> str) where T : IEquatable<T>;
