+++ FuzzySharp/SimilarityRatio/Scorer/StrategySensitive/Simple/PartialRatioScorer.cs
@@ -5,6 +5,6 @@ namespace Raffinert.FuzzySharp.SimilarityRatio.Scorer.StrategySensitive
 {
     public class PartialRatioScorer : SimpleRatioScorerBase
     {
-        protected override Func<string, string, int> Scorer => PartialRatioStrategy.Calculate;
+        protected override FuzzySharp.Scorer Scorer => PartialRatioStrategy.Calculate;
     }
 }
