+++ FuzzySharp/SimilarityRatio/Scorer/StrategySensitive/StrategySensitiveScorerBase.cs
@@ -4,6 +4,6 @@ namespace Raffinert.FuzzySharp.SimilarityRatio.Scorer.StrategySensitive
 {
     public abstract class StrategySensitiveScorerBase : ScorerBase
     {
-        protected abstract Func<string, string, int> Scorer { get; }
+        protected abstract FuzzySharp.Scorer Scorer { get; }
     }
 }
