+++ FuzzySharp/SimilarityRatio/Scorer/StrategySensitive/TokenSet/PartialTokenSetScorer.cs
@@ -5,6 +5,6 @@ namespace Raffinert.FuzzySharp.SimilarityRatio.Scorer.StrategySensitive
 {
     public class PartialTokenSetScorer : TokenSetScorerBase
     {
-        protected override Func<string, string, int> Scorer => PartialRatioStrategy.Calculate;
+        protected override FuzzySharp.Scorer Scorer => PartialRatioStrategy.Calculate;
     }
 }
