+++ FuzzySharp/SimilarityRatio/Strategy/DefaultRatioStrategy.cs
@@ -10,7 +10,7 @@ namespace Raffinert.FuzzySharp.SimilarityRatio.Strategy
             {
                 return 0;
             }
-            return (int)Math.Round(100 * NewLevenshtein.GetRatio(input1, input2));
+            return (int)Math.Round(100 * Levenshtein.GetRatio(input1, input2));
         }
     }
 }
